+++ packages/components/SignaturePad/index.ts
@@ -1,0 +1,2 @@
+export { SignaturePad } from './SignaturePad'
+export { SignaturePadDemo } from './SignaturePadDemo'