+++ src/pages/index/demo/components/signature-pad.tsx
@@ -1,0 +1,3 @@
+import { ZaSignaturePadDemo } from '@zealous-admin/components/index'
+
+export default ZaSignaturePadDemo